Final testing version

+++ sia-admin/chk_install.php
@@ -8,14 +8,13 @@
 <body class="login-index">
 <?php
 require("constants.php");
-if(mysqli_connect_errno() >0){
-	unlink("../sia-config.php");
+if(mysqli_connect_errno() >0 || check_install_complete() == false){
+	if(check_install_complete()){ unlink("../sia-config.php"); }
+	
 ?>
-
-
-
 <div class="add-data-basefull-width-error">
 	<h1>Error establishing a database connection</h1>
+	<hr></hr>
 	<p>This either means that the username and password information in your <code>sia-config.php</code> file is incorrect or we can't contact the database server at <code>localhost</code>. This could mean your host's database server is down.</p>
 	<ul>
 		<li>Are you sure you have the correct username and password?</li>
@@ -28,11 +27,26 @@ if(mysqli_connect_errno() >0){
 	
 </div>
 <?php
-}else{
+}/*else if($sia_obj->chk_tables() == false){
+?>
+	<div class="add-data-basefull-width-error">
+		<h1 class="screen-reader-text">Database Error</h1>
+		<hr></hr>
+		<p>A table is missing in your data base. You can fix it manually or Slack will do this for you.</p>
+		<p class="step"><a href="chk_install.php?fix_table=yes" class="button button-large">Click Here</a></p>
+		<?php if(isset($_GET['fix_table']) && !empty($_GET['fix_table']) && $_GET['fix_table'] == "yes"){
+			$sia_obj->create_basic_set_up();
+			header("Location: index.php");
+		}
+		?>
+	</div>
+<?php
+}*/else{
 	$run = $sia_obj->create_basic_set_up();
 ?>
 <div class="add-data-basefull-width-error">
 	<h1 class="screen-reader-text">Successful database connection</h1>
+	<hr></hr>
 	<p>All right, sparky! You've made it through this part of the installation. Slack Invite can now communicate with your database. If you are ready, time now to...</p>
 	<p class="step"><a href="create_admin.php" class="button button-large">Run the install</a></p>
 </div>
